Add line-buffered mode
The stream-buffer will read as much data as it can, and output a line
for each read(2). When the buffer becomes full, it is dumped and the
process begins anew.

This means that the output produces is something like this: a line, a
line, a line, a BIG CHUNK of data, a line, a line, etc. Not optimal, so
to say. The strategy will surely change in later versions, but now teh
basics work and are kinda good enough for my needs.