Fix buffer hanging if pipe is not readable on exit
The stream-buffer attempted to read whetever was in the pipe before it
drained the buffer for the last time to avoid losing data. However, this
proved to be more troublesome than helpful.

It turned out that if both the program whose standard output was being
buffered and the buffer itself received SIGTERM in short sequence the
buffer would hang. The program was no longer producing output so the
buffer had nothing to read, and yet it was trying to. Without the pipe's
file descriptor set to non-blocking the stream got stuck indefinitely.